Collision detection added to the main program
Just for testing at this stage. When camera collides a building, the
label "Collision" on the webpage becomes true. Since I use a unit cube
as the bounding volume of the camera, it's a little bit hard to go
through some narrow gaps between buildings without collision. Due to the
near clipping plane, you may think the detection is delayed because the
building is clipped before the camera actually hits it.